Revert "Unit tests implementation  + minor bug fix"
