+++ ListMate/UI/Tabs/List/Main/Controller/ListViewController.swift
@@ -10,8 +10,7 @@ import UIKit
 class ListViewController: BaseViewController {
     
     private lazy var viewModel: ListViewModel = {
-        let manager = DataManager()
-        let model = ListViewModel(session: ProductSession.shared, manager: manager)
+        let model = ListViewModel(session: .shared)
         model.delegate = self
         return model
     }()
